requirements.txt
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
fonttools==4.39.0
idna==3.4
ifaddr==0.2.0
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
jupyter_client==8.0.3
jupyter_core==5.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.7.1
pyusb==1.2.1
PyVISA==1.13.0
PyVISA-py==0.6.3
pyzmq==25.0.1
requests==2.28.2
scipy==1.10.1
six==1.16.0
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
zeroconf==0.47.3
